requirements.txt
psycopg2 
streamlit 
matplotlib 
pandas 
sqlalchemy
seaborn
python-dotenv